Move jani_variable generator to jani_utils
