+++ src/as2fm/jani_generator/jani_entries/jani_utils.py
@@ -17,8 +17,9 @@
 
 from typing import Tuple, Type, get_args
 
-from as2fm.as2fm_common.common import is_array_type
+from as2fm.as2fm_common.common import ValidTypes, get_default_expression_for_type, is_array_type
 from as2fm.jani_generator.jani_entries import JaniExpression, JaniVariable
+from as2fm.jani_generator.jani_entries.jani_expression_generator import array_create_operator
 
 
 def is_expression_array(expr: JaniExpression) -> bool:
@@ -57,3 +58,14 @@ def get_array_variable_info(jani_var: JaniVariable) -> Tuple[Type, int]:
     ), f"Unsupported array type {array_type} found in JANI variable {jani_var.name()}."
     max_size = get_expression_array_length(jani_var.get_init_expr())
     return (array_type, max_size)
+
+
+def generate_jani_variable(var_name: str, var_type: ValidTypes, array_size: int):
+    """Helper to make a JaniVariable object."""
+    # TODO: Move it to jani_utils.py
+    if is_array_type(var_type):
+        array_type = get_args(var_type)[0]
+        init_value = array_create_operator("__array_iterator", array_size, array_type(0))
+    else:
+        init_value = JaniExpression(get_default_expression_for_type(var_type))
+    return JaniVariable(var_name, var_type, init_value)
